feat: add `onError` option
The `queueSend` option cannot intercept the `send` promise anymore. The `onError` callback can be used to rethrow the error and/or override the default error logging. If it rethrows the error, the original caller will have its promise rejected. For this reason, you should only rethrow when the `method` argument of `onError` equals "setUserProps", since that is the only client method that returns its promise. All other methods are considered non-critical, and so cannot be waited on.